py3/pyproject.toml
[tool.poetry]
name = "highload"
version = "0.1.0"
description = "highload research project"
authors = ["VladNF <j@VladNF.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
yoyo-migrations = "^8.2.0"
psycopg2-binary = "^2.9.5"
environs = "^9.5.0"
asyncpg = "^0.27.0"
pyjwt = "^2.6.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^1.10.4"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
fastapi-code-generator = "^0.4.0"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
